Fix robot reference in constructor
Prior to this commit the constructor for the class `Listener` in
conversation.coffee referenced a local variable `robot` which isn't defined in
this context. I believe perhaps the intention was to reference the `@robot`
argument passed to the constructor as is done later in the `call` function?
This commit updates on this assumption - replacing `robot` with `@robot`.

Signed-off-by: Moses Mendoza <mendoza.moses@gmail.com>